pylint refactor
