+++ carbon-rating.py
@@ -90,7 +90,7 @@ def main(argv):
             elif (file_ending == ".xml"):  # https://example.com/sitemap.xml
                 from engines.sitemap import read_sites, add_site, delete_site
             else:
-                from engines.json import read_tests, read_sites, add_site, delete_site
+                from engines.json_engine import read_tests, read_sites, add_site, delete_site
             pass
         elif opt in ("-o", "--output"):  # output file path
             output_filename = arg
